refactor(mcp): decompose maestro-server, validate-plan, and rebuild-frontmatter
Wave 2 of the complexity-reduction plan. Behavior-preserving: zero
generator drift, all 851 tests pass.

- Split maestro-server.js (398 lines) into three single-purpose core
  modules and a 115-line composer:
    * core/line-reader.js    - readline stdin -> JSON dispatch
    * core/protocol-dispatcher.js - JSON-RPC 2.0 routing, result builders
    * core/project-root-cache.js  - workspace env + roots/list caching
  Public exports (DEFAULT_PROTOCOL_VERSION, SERVER_INFO,
  createInitializeResult, createToolErrorResult, createToolSuccessResult,
  normalizeRuntimeConfig, runRuntimeServer, main) are preserved.

- Split validate-plan.js (341 lines) into four orthogonal checkers plus a
  65-line orchestrator:
    * validation/schema-checker.js       - shape, phase_count, duplicate_id,
                                           dangling_dependency
    * validation/agent-checker.js        - unknown_agent, agent_capability_mismatch
    * validation/dag-checker.js          - cyclic_dependency, redundant_dependency,
                                           depths, parallelization profile
    * validation/file-overlap-checker.js - file_overlap on parallel batches
  Violation ordering and error codes preserved; handler's public contract
  unchanged.

- Rewrite rebuild-frontmatter.js as table-driven: per-field emitter map plus
  per-runtime ordered field lists. Claude's block-scalar description and
  color emission, gemini's tools-before-turns layout, and the default
  (codex/qwen) shape are now data rather than conditionals. Generator output
  is byte-identical.
